requirements/test.txt
nose==1.3.7
requests==2.13.0
pyyaml==3.12